+++ source/testing/CXXExecutableModel.h
@@ -83,17 +83,17 @@ public:
      * @param[out] values an array of at least length len which will store the
      *                returned boundary species amounts.
      */
-    virtual int getFloatingSpeciesAmounts(int len, int const *indx,
+    virtual int getFloatingSpeciesAmounts(size_t len, int const *indx,
             double *values);
 
-    virtual int setFloatingSpeciesAmounts(int len, int const *indx,
+    virtual int setFloatingSpeciesAmounts(size_t len, int const *indx,
             const double *values);
 
-    virtual int getFloatingSpeciesAmountRates(int len, int const *indx,
+    virtual int getFloatingSpeciesAmountRates(size_t len, int const *indx,
             double *values);
 
 
-    virtual int getFloatingSpeciesConcentrationRates(int len, int const *indx,
+    virtual int getFloatingSpeciesConcentrationRates(size_t len, int const *indx,
                 double *values);
 
     /**
@@ -104,7 +104,7 @@ public:
      * @param[out] values an array of at least length len which will store the
      *                returned boundary species amounts.
      */
-    virtual int getFloatingSpeciesConcentrations(int len, int const *indx,
+    virtual int getFloatingSpeciesConcentrations(size_t len, int const *indx,
             double *values);
 
     /**
@@ -115,7 +115,7 @@ public:
      * @param[in] values an array of at least length len which will store the
      *                returned boundary species amounts.
      */
-    virtual int setFloatingSpeciesConcentrations(int len, int const *indx,
+    virtual int setFloatingSpeciesConcentrations(size_t len, int const *indx,
             double const *values);
 
     /**
@@ -128,14 +128,14 @@ public:
      * updating the values on which it depends, it can not be set
      * directly.
      */
-    virtual int setFloatingSpeciesInitConcentrations(int len, int const *indx,
+    virtual int setFloatingSpeciesInitConcentrations(size_t len, int const *indx,
                 double const *values);
 
     /**
      * Get the initial concentrations of the floating species,
      * uses the same indexing as the other floating species methods.
      */
-    virtual int getFloatingSpeciesInitConcentrations(int len, int const *indx,
+    virtual int getFloatingSpeciesInitConcentrations(size_t len, int const *indx,
                     double *values);
 
     /**
@@ -148,14 +148,14 @@ public:
      * updating the values on which it depends, it can not be set
      * directly.
      */
-    virtual int setFloatingSpeciesInitAmounts(int len, int const *indx,
+    virtual int setFloatingSpeciesInitAmounts(size_t len, int const *indx,
                 double const *values);
 
     /**
      * Get the initial amounts of the floating species,
      * uses the same indexing as the other floating species methods.
      */
-    virtual int getFloatingSpeciesInitAmounts(int len, int const *indx,
+    virtual int getFloatingSpeciesInitAmounts(size_t len, int const *indx,
                     double *values);
 
     /************************ End Floating Species Section ************************/
@@ -184,7 +184,7 @@ public:
      * @param[out] values an array of at least length len which will store the
      *                returned boundary species amounts.
      */
-    virtual int getBoundarySpeciesAmounts(int len, int const *indx,
+    virtual int getBoundarySpeciesAmounts(size_t len, int const *indx,
             double *values);
 
 
@@ -196,7 +196,7 @@ public:
      * @param[out] values an array of at least length len which will store the
      *                returned boundary species amounts.
      */
-    virtual int getBoundarySpeciesConcentrations(int len, int const *indx,
+    virtual int getBoundarySpeciesConcentrations(size_t len, int const *indx,
             double *values);
 
     /**
@@ -207,7 +207,7 @@ public:
      * @param[in] values an array of at least length len which will store the
      *                returned boundary species amounts.
      */
-    virtual int setBoundarySpeciesConcentrations(int len, int const *indx,
+    virtual int setBoundarySpeciesConcentrations(size_t len, int const *indx,
             double const *values);
 
 
@@ -243,10 +243,10 @@ public:
      * @param[out] values an array of at least length len which will store the
      *                returned boundary species amounts.
      */
-    virtual int getGlobalParameterValues(int len, int const *indx,
+    virtual int getGlobalParameterValues(size_t len, int const *indx,
             double *values);
 
-    virtual int setGlobalParameterValues(int len, int const *indx,
+    virtual int setGlobalParameterValues(size_t len, int const *indx,
             const double *values);
 
 
@@ -271,10 +271,10 @@ public:
      * @param[out] values an array of at least length len which will store the
      *                returned boundary species amounts.
      */
-    virtual int getCompartmentVolumes(int len, int const *indx,
+    virtual int getCompartmentVolumes(size_t len, int const *indx,
             double *values);
 
-    virtual int setCompartmentVolumes(int len, int const *indx,
+    virtual int setCompartmentVolumes(size_t len, int const *indx,
                 const double *values);
 
     /**
@@ -287,14 +287,14 @@ public:
      * updating the values on which it depends, it can not be set
      * directly.
      */
-    virtual int setCompartmentInitVolumes(int len, int const *indx,
+    virtual int setCompartmentInitVolumes(size_t len, int const *indx,
                 double const *values);
 
     /**
      * Get the initial volume of the compartments,
      * uses the same indexing as the other compartment methods.
      */
-    virtual int getCompartmentInitVolumes(int len, int const *indx,
+    virtual int getCompartmentInitVolumes(size_t len, int const *indx,
                     double *values);
 
 
@@ -361,8 +361,8 @@ public:
     virtual int getNumConservedMoieties();
     virtual int getConservedMoietyIndex(const std::string& eid);
     virtual std::string getConservedMoietyId(int index);
-    virtual int getConservedMoietyValues(int len, int const *indx, double *values);
-    virtual int setConservedMoietyValues(int len, int const *indx,
+    virtual int getConservedMoietyValues(size_t len, int const *indx, double *values);
+    virtual int setConservedMoietyValues(size_t len, int const *indx,
             const double *values);
 
     virtual int getNumRateRules();
@@ -381,7 +381,7 @@ public:
     /**
      * get the name of the specified reaction
      */
-    virtual std::string getReactionId(int index);
+    virtual std::string getReactionId(size_t index);
 
     /**
      * get the vector of reaction rates.
@@ -391,7 +391,7 @@ public:
      * reaction rates are copied directly into the suplied buffer.
      * @param values: pointer to user suplied buffer where rates will be stored.
      */
-    virtual int getReactionRates(int len, int const *indx,
+    virtual int getReactionRates(size_t len, int const *indx,
                 double *values);
 
     /**
@@ -478,7 +478,7 @@ public:
      * So, on every modern system I'm aware of, bool is an unsigned char, so
      * use that data type here.
      */
-    virtual int getEventTriggers(int len, const int *indx, unsigned char *values);
+    virtual int getEventTriggers(size_t len, const int *indx, unsigned char *values);
 
     virtual int applyEvents(double timeEnd, const unsigned char* previousEventStatus,
                 const double *initialState, double* finalState);
